refactor: sensor state logic (#161)
Use the same state logic as for the other entities.
I had a non-reproducible issue where suddenly all sensor entities
remained in UNAVAILABLE state, whereas the remote- and media-player
entities showed the correct state when turning the AVR on and off.

Add a safeguard by setting all entities to UNAVAILABLE if the AVR
connection fails.